echoai-backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy==2.0.41
pydantic==2.7.4
python-jose[cryptography]==3.3.0
passlib==1.7.4
python-dotenv==1.0.1
